Moved splashscreen to createAccount routing
