+++ src/app/create-account/carousel-grid.scss
