+++ src/app/create-account/create-account-routing.module.ts
@@ -2,11 +2,11 @@ import { NgModule } from "@angular/core";
 import { Routes } from "@angular/router";
 import { NativeScriptRouterModule } from "nativescript-angular/router";
 
-import { SplashscreenComponent } from "./splashscreen.component";
+import { IndexComponent } from "./index/index.component";
 import { GenerateScreenComponent } from "./generate/generate.component";
 
 const routes: Routes = [
-  { path: '', component: SplashscreenComponent },
+  { path: '', component: IndexComponent },
   { path: 'generate', component: GenerateScreenComponent }
 ];
 
@@ -18,4 +18,4 @@ const routes: Routes = [
     NativeScriptRouterModule
   ]
 })
-export class SplashscreenRoutingModule { }
+export class CreateAccountRoutingModule { }
