--- src/app/create-account/create-account.component.scss
@@ -1,114 +1,0 @@
-// .page {
-//   background-color: blue;
-// }
-
-
-@keyframes fingerSwipe {
-  0% {
-    transform: translate3d(0, 20%, 0) rotate(20) scale(1.2);
-  }
-  10% {
-    transform: translate3d(0, 0, 0) rotate(20) scale(1);
-  }
-  50% {
-    transform: translate3d(70%, 0, 0) rotate(-20) scale(1);
-  }
-  90% {
-    transform: translate3d(0, 0, 0) rotate(20) scale(1);
-  }
-  100% {
-    transform: translate3d(0, 0, 0) rotate(20) scale(1.2);
-  }
-}
-
-.swipeFinger {
-  transform: translate3d(0, 20%, 0) rotate(20) scale(1.2);
-
-  animation-name: fingerSwipe;
-  animation-duration: 5s;
-  animation-fill-mode: both;
-  animation-iteration-count: infinite;
-  animation-timing-function: linear;
-}
-
-.h1 {
-  margin-top: 30;
-  margin-bottom: 30;
-  color: #050709;
-}
-
-.card {
-  justify-self: center;
-  align-self: center;
-}
-
-.seedPool {
-  font-size: 13;
-}
-
-.agreements-wrap {
-  text-align: center;
-  width: 90%;
-  padding-top: 15;
-  padding-bottom: 30;
-
-  > Label {
-    margin-bottom: 10;
-  }
-
-  .agreements {
-    justify-content: space-around;
-  }
-}
-
-.registered--view {
-  flex-direction: column;
-}
-
-.waves {
-  justify-content: center;
-  align-content: center;
-  align-items: center;
-  margin-top: 40;
-  flex-direction: row;
-  opacity: 0.4;
-}
-
-
-.header-wrap {
-  flex-direction: column;
-
-  .header-bar {
-    text-align: center;
-    padding-top: 20;
-    font-size: 24;
-    font-weight: bold;
-  }
-}
-
-Label {
-  word-wrap: break-word;
-  flex-wrap: wrap;
-  overflow-wrap: break-word;
-}
-
-.seedPool {
-  min-height: 200;
-}
-
-.actions {
-  margin-top: 20;
-  margin-bottom: 20;
-  width: 90%;
-  
-  Label {
-    text-align: center;
-    align-self: flex-end;
-  }
-
-  Button {
-    width: 100%;
-  }
-}
-
-
