+++ src/app/create-account/generate/generate.component.ts
