+++ src/app/create-account/wallet/wallet.component.scss
